refactor(server): colocate runtime services by domain
