+++ packages/server/src/__tests__/admin-disconnect.test.ts
@@ -1,7 +1,7 @@
 import type { FastifyInstance } from "fastify";
 import { describe, expect, it } from "vitest";
 import { createAgent } from "../services/agents/identity.js";
-import * as presenceService from "../services/presence.js";
+import * as presenceService from "../services/runtime/presence.js";
 import { createAdminContext, useTestApp } from "./helpers.js";
 
 describe("Admin Agent Disconnect API", () => {
