+++ packages/server/src/__tests__/client-service.test.ts
@@ -4,9 +4,9 @@ import { agentPresence } from "../db/schema/agent-presence.js";
 import { agents } from "../db/schema/agents.js";
 import { clients } from "../db/schema/clients.js";
 import { createAgent, suspendAgent } from "../services/agents/identity.js";
-import * as clientService from "../services/client.js";
-import * as presenceService from "../services/presence.js";
-import { recordClientHeartbeat } from "../services/runtime-liveness.js";
+import * as clientService from "../services/runtime/client.js";
+import { recordClientHeartbeat } from "../services/runtime/liveness.js";
+import * as presenceService from "../services/runtime/presence.js";
 import { createTestAgent, useTestApp } from "./helpers.js";
 
 /**
