+++ packages/server/src/__tests__/me-leave-managed-agents.test.ts
@@ -5,10 +5,10 @@ import { describe, expect, it } from "vitest";
 import { agents as agentsTable } from "../db/schema/agents.js";
 import { members as membersTable } from "../db/schema/members.js";
 import { createAgent, updateAgent } from "../services/agents/identity.js";
-import { retireClient } from "../services/client.js";
 import { createMember } from "../services/member.js";
 import { leaveOrganization } from "../services/membership.js";
 import { createOrganization } from "../services/organization.js";
+import { retireClient } from "../services/runtime/client.js";
 import { seedClient, useTestApp } from "./helpers.js";
 
 /**
