+++ packages/server/src/api/agent/me.ts
@@ -1,7 +1,7 @@
 import type { FastifyInstance } from "fastify";
 import { requireAgent } from "../../middleware/require-identity.js";
 import * as agentService from "../../services/agents/identity.js";
-import * as presenceService from "../../services/presence.js";
+import * as presenceService from "../../services/runtime/presence.js";
 
 export async function agentMeRoutes(app: FastifyInstance): Promise<void> {
   app.get("/me", async (request) => {
