+++ packages/server/src/api/agent/ws-client/client-frames.ts
@@ -10,11 +10,11 @@ import type { WebSocket } from "ws";
 import { members } from "../../../db/schema/members.js";
 import { ClientOrgMismatchError, ClientRetiredError, ClientUserMismatchError } from "../../../errors.js";
 import { setWsConnectionAttrs } from "../../../observability/index.js";
-import * as clientService from "../../../services/client.js";
-import * as connectionManager from "../../../services/connection-manager.js";
 import type { Notifier } from "../../../services/notifier.js";
-import { storeModelCatalogRpcResult } from "../../../services/provider-models-rpc.js";
-import * as runtimeLivenessService from "../../../services/runtime-liveness.js";
+import * as clientService from "../../../services/runtime/client.js";
+import * as connectionManager from "../../../services/runtime/connection-manager.js";
+import * as runtimeLivenessService from "../../../services/runtime/liveness.js";
+import { storeModelCatalogRpcResult } from "../../../services/runtime/rpc/provider-models.js";
 import { setAuthWsAttrs } from "./auth.js";
 import type { ClientWsConnectionContext } from "./connection-context.js";
 import type { InboxDeliveryCoordinator } from "./inbox-delivery.js";
