+++ packages/server/src/api/agent/ws-client/session-frames.ts
@@ -15,13 +15,16 @@ import type { WebSocket } from "ws";
 import { agentChatSessions } from "../../../db/schema/agent-chat-sessions.js";
 import * as activityService from "../../../services/chat/sessions/activity.js";
 import * as sessionEventService from "../../../services/chat/sessions/events.js";
-import * as connectionManager from "../../../services/connection-manager.js";
 import * as contextTreeIoService from "../../../services/context-tree/io.js";
 import * as landingCampaignChatStateService from "../../../services/landing-campaigns/chat-state.js";
 import * as notificationService from "../../../services/notification.js";
 import type { Notifier } from "../../../services/notifier.js";
-import * as presenceService from "../../../services/presence.js";
-import { type SessionCommandRpcPhase, storeSessionCommandRpcResult } from "../../../services/session-command-rpc.js";
+import * as connectionManager from "../../../services/runtime/connection-manager.js";
+import * as presenceService from "../../../services/runtime/presence.js";
+import {
+  type SessionCommandRpcPhase,
+  storeSessionCommandRpcResult,
+} from "../../../services/runtime/rpc/session-command.js";
 import { KeyedOperationQueue } from "../../../utils/keyed-operation-queue.js";
 import type { ClientWsConnectionContext } from "./connection-context.js";
