+++ packages/server/src/api/orgs/agents.ts
@@ -11,8 +11,8 @@ import { BadRequestError, ForbiddenError } from "../../errors.js";
 import { requireOrgMembership } from "../../scope/require-org.js";
 import * as agentService from "../../services/agents/identity.js";
 import { resolveAvatarImageUrl } from "../../services/agents/identity.js";
-import { sendToClient } from "../../services/connection-manager.js";
 import { assertMetadataDoesNotClaimLandingCampaignTrial } from "../../services/landing-campaigns/guards.js";
+import { sendToClient } from "../../services/runtime/connection-manager.js";
 
 function serializeNewChatDefaultCandidate(agent: agentService.NewChatDefaultCandidateAgent) {
   return { ...agent, createdAt: agent.createdAt.toISOString() };
