+++ packages/server/src/services/membership.ts
@@ -11,8 +11,8 @@ import { ConflictError, NotFoundError } from "../errors.js";
 import { uuidv7 } from "../uuid.js";
 import { lockWatcherProjectionMemberMutation } from "./chat/membership/lock.js";
 import { lockWatcherProjectionForMemberChanges, recomputeWatcherChats } from "./chat/membership/watcher.js";
-import { forceDisconnect } from "./connection-manager.js";
-import * as presenceService from "./presence.js";
+import { forceDisconnect } from "./runtime/connection-manager.js";
+import * as presenceService from "./runtime/presence.js";
 import { suspendGitlabLinksForMembership } from "./scm/gitlab/identities.js";
 
 /**
